feat: Rename Alertmanager 'rule' label to standard 'alertname'
What type of PR is this?

/kind feature

Any specific area of the project related to this PR?

/area outputs

What this PR does / why we need it:
Alertmanager doesn't require a label called 'alertname', but it is a universally expected one which all alerts originating from Prometheus itself contain. Without it downstream grouping/routing/templating can require some customisation in order to handle alerts from falcosidekick.

Signed-off-by: Ed Kennard <ed@kennard.net>